+++ src/main/java/coffeemeet/server/common/config/AuthWebConfig.java
@@ -2,7 +2,7 @@ package coffeemeet.server.common.config;
 
 import coffeemeet.server.auth.domain.JwtTokenProvider;
 import coffeemeet.server.auth.implement.RefreshTokenQuery;
-import coffeemeet.server.common.resolver.UserArgumentResolver;
+import coffeemeet.server.common.presentation.resolver.UserArgumentResolver;
 import coffeemeet.server.oauth.utils.converter.OAuthProviderConverter;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
