+++ src/test/java/coffeemeet/server/certification/service/CertificationServiceTest.java
@@ -18,8 +18,8 @@ import static org.mockito.Mockito.only;
 import coffeemeet.server.certification.implement.CertificationCommand;
 import coffeemeet.server.certification.implement.EmailVerificationCommand;
 import coffeemeet.server.certification.implement.EmailVerificationQuery;
-import coffeemeet.server.common.media.EmailSender;
-import coffeemeet.server.common.media.MediaManager;
+import coffeemeet.server.common.implement.EmailSender;
+import coffeemeet.server.common.implement.MediaManager;
 import coffeemeet.server.common.util.FileUtils;
 import coffeemeet.server.user.domain.User;
 import coffeemeet.server.user.implement.UserQuery;
